+++ lib/layrz_models.dart
@@ -6,7 +6,6 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart' hide Action;
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:layrz_icons/layrz_icons.dart';
-import 'package:layrz_models/utils/utils.dart';
 
 import 'src/actions/actions.dart';
 // Internal imports
@@ -34,6 +33,8 @@ import 'src/sensors/sensors.dart';
 import 'src/users/users.dart';
 import 'src/vision/vision.dart';
 import 'src/workspace/workspace.dart';
+import 'src/triggers/triggers.dart';
+import 'src/care_protocols/care_protocols.dart';
 
 export 'src/actions/actions.dart';
 // Internal exports
@@ -69,6 +70,8 @@ export 'src/tenvio/tenvio.dart';
 export 'src/users/users.dart';
 export 'src/vision/vision.dart';
 export 'src/workspace/workspace.dart';
+export 'src/triggers/triggers.dart';
+export 'src/care_protocols/care_protocols.dart';
 
 // Freezed
 part 'layrz_models.freezed.dart';
@@ -79,7 +82,6 @@ part 'src/accessibility/shortcut.dart';
 part 'src/algorithm.dart';
 part 'src/avatar.dart';
 part 'src/billing_plan.dart';
-part 'src/care_protocol.dart';
 part 'src/case.dart';
 part 'src/category.dart';
 part 'src/checkpoint.dart';
@@ -105,7 +107,6 @@ part 'src/tag.dart';
 part 'src/telemetry.dart';
 part 'src/timezone.dart';
 part 'src/token.dart';
-part 'src/trigger.dart';
 part 'src/trip.dart';
 part 'src/undeclarated_models.dart';
 part 'src/weekdays.dart';