+++ lib/src/care_protocols/care_protocols.dart
@@ -1,0 +1,20 @@
+library;
+
+import 'package:collection/collection.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:layrz_models/layrz_models.dart';
+
+// Freezed
+part 'care_protocols.freezed.dart';
+part 'care_protocols.g.dart';
+
+// Enums
+part 'src/mode.dart';
+part 'src/answer_kind.dart';
+
+// Models
+part 'src/care_protocol.dart';
+part 'src/care_task.dart';
+
+// Legacy Converters
+part 'src/legacy_converters.dart';